1.支持读取银行卡硬件操作. 2.修改DecToHex函数bug.

+++ ETtradeClient/project/ETradeClient/hardware/dili_card_device.cpp
@@ -17,7 +17,7 @@ std::string DecToHex(T d, bool show_base = false)
 	std::stringstream ss;
 	if (show_base)
 		ss << "0x";
-	ss << std::setfill('0') << std::setw(sizeof(T) * 2) << std::hex << d;
+	ss << std::setfill('0') << std::setw(sizeof(T) * 2) << std::hex << static_cast<int>(d);
 	return ss.str();
 }
